model/model.py
import torch
import torch.nn as nn
from ultralytics import YOLO


class HumanActionModel(nn.Module):
    def __init__(self, num_actions=4, detection_model_name='yolov8n.pt'):
        super(HumanActionModel, self).__init__()
        # Load a pre-trained YOLO model for object detection
        # We'll use this as a feature extractor or for initial human detection.
        self.detector = YOLO(detection_model_name)

        # Freeze detector layers if you only want to train the classification head
        # for param in self.detector.model.parameters():
        #     param.requires_grad = False

        # Example: Get the number of output features from the detector's backbone
        # This part is highly dependent on the specific YOLO model structure
        # and how you intend to add the classification head.
        # For simplicity, let's assume we extract features after detection.

        # Placeholder for an action classification head
        # This head would take features from detected human bounding boxes
        # and classify the action.
        # The actual input size will depend on how features are extracted.
        # For example, if using RoIAlign and a small CNN or MLP:
        # self.action_classifier = nn.Sequential(
        #     nn.Linear(in_features=SOME_FEATURE_SIZE, out_features=256),
        #     nn.ReLU(),
        #     nn.Dropout(0.5),
        #     nn.Linear(in_features=256, out_features=num_actions)
        # )

        # For now, we'll focus on using YOLO for detection.
        # The strategy for action classification needs to be refined:
        # 1. Train YOLO to detect humans (1 class). Then, for each detected human,
        #    crop the image and pass it to a separate action classifier.
        # 2. Modify YOLO to have `num_actions` output classes directly, where each class
        #    represents an action (e.g., standing_person, sitting_person). This requires
        #    reformatting labels and retraining YOLO.
        # 3. Use YOLO for detection and add a custom classification head that operates
        #    on features from the detected regions.

        print(
            f"HumanActionModel initialized with YOLO: {detection_model_name}")
        print(f"Number of action classes: {num_actions}")
        print(
            "Note: Action classification head needs to be properly defined and integrated.")

    def forward(self, images):
        # images: a batch of images (e.g., torch.Tensor of shape [B, C, H, W])

        # Perform detection using the YOLO model
        # The output format depends on the Ultralytics YOLO version/API.
        # Typically, it returns a list of Results objects, one per image.
        detection_results = self.detector(images)

        # Post-processing:
        # For each image in the batch:
        #   For each detected human:
        #     - Extract the bounding box.
        #     - Crop the human region from the image (or extract features from this region).
        #     - Pass the cropped region/features to the action classifier.

        # This is a placeholder for the combined output.
        # The actual implementation will depend on the chosen strategy for action classification.
        # For now, let's just return the detection results.
        # The action classification logic will be added later.

        # Example (conceptual):
        # all_actions_logits = []
        # for i, result in enumerate(detection_results):
        #     image_actions_logits = []
        #     for box in result.boxes: # Assuming result.boxes contains detected bounding boxes
        #         if int(box.cls) == 0: # Assuming class 0 is 'human' if YOLO is trained for humans
        #             # Crop image or extract features for this box
        #             # x1, y1, x2, y2 = box.xyxy[0]
        #             # cropped_human = images[i][:, int(y1):int(y2), int(x1):int(x2)]
        #             # features = self.feature_extractor_for_action(cropped_human)
        #             # action_logits = self.action_classifier(features)
        #             # image_actions_logits.append(action_logits)
        #             pass # Placeholder
        #     all_actions_logits.append(image_actions_logits)

        # return detection_results, all_actions_logits # Or a more structured output

        return detection_results  # Returning raw YOLO results for now


if __name__ == '__main__':
    # Example usage:
    model = HumanActionModel(num_actions=4)

    # Create a dummy input image (batch of 1, 3 channels, 640x345)
    # Note: YOLO might internally resize/preprocess, check its docs.
    # The input size 640x345 is mentioned in the requirements.
    # Ultralytics YOLO typically expects images in a certain range (e.g. 640).
    # It handles padding/resizing.
    dummy_image = torch.randn(1, 3, 345, 640)

    print(f"Dummy input shape: {dummy_image.shape}")

    # Put model in eval mode for inference if not training
    model.eval()

    with torch.no_grad():
        try:
            results = model(dummy_image)
            # 'results' will be a list of Ultralytics Results objects
            # Each Results object contains boxes, masks, probs, etc.
            if results:
                print(
                    f"Model executed. Number of results (images processed): {len(results)}")
                for i, r in enumerate(results):
                    print(f"--- Image {i+1} ---")
                    print(f"  Detected boxes: {len(r.boxes)}")
                    if len(r.boxes) > 0:
                        print(f"  Example box (xyxy): {r.boxes[0].xyxy}")
                        print(f"  Example box conf: {r.boxes[0].conf}")
                        print(f"  Example box cls: {r.boxes[0].cls}")
            else:
                print(
                    "Model executed, but no results returned (this might be unexpected).")

        except Exception as e:
            print(f"Error during model execution: {e}")
            import traceback
            traceback.print_exc()

    print("Model definition and basic test completed.")
    print("Further work: Integrate action classification head, define feature extraction, and adapt training.")
